User registration and login in a single Auth controller
