+++ drboson-backend/src/main/java/com/filipovski/drboson/drboson/web/controllers/UserController.java
@@ -3,11 +3,12 @@ package com.filipovski.drboson.drboson.web.controllers;
 import com.filipovski.drboson.drboson.model.User;
 import com.filipovski.drboson.drboson.repository.UserRepository;
 import com.filipovski.drboson.drboson.service.dtos.UserProjection;
-import org.apache.jena.tdb.lib.StringAbbrev;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
-import org.springframework.util.MimeTypeUtils;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 
 @RestController
